Wrapper script
Having the wrapper script as the parent process makes it easier to
detect if the `cloud_sql_proxy` processes fails and makes it possible to
force a restart of the dyno. It also makes it easier to delay starting
the app while `cloud_sql_proxy` is booting.

Adapted from [heroku/heroku-buildpack-nginx][1]. The only meaningful
change was to remove the non-zero exit code. Our wrapper script will
probably be used as part of a release phase and returning a non-zero
code would cause the release command to fail.

[1]: https://github.com/heroku/heroku-buildpack-nginx